requirements.txt
blessings==1.6
click==4.1
Flask==0.10.1
Flask-SQLAlchemy==2.0
gunicorn==19.3.0
lxml==3.4.4
path.py==7.4
psycopg2==2.6.1
q==2.6
requests==2.7.0
Scrapy==1.0.1
service-identity==14.0.0
SQLAlchemy==1.0.8
